refactor(app): separate route files from library modules
app/ held 23 files of which 15 were not routes: the server action, the
caches, the rate limiter, the logger, the types and the rest sat at the
same level as page.tsx and layout.tsx, so listing the directory that
defines what the app serves mostly showed things it does not serve.
`app/logger.ts` was the clearest case -- imported by a route as
`@/app/logger`, a module crossing route boundaries from inside the
routes directory.

Ten modules and their tests moved to app/_lib/. The issue offered `lib/`
at the root as the alternative, but #143 had just deleted the root
components/ directory in favour of colocation under app/, and a root
lib/ would reverse that a day later. Underscore-prefixed folders under
app/ are now the one rule: _home for the feature, _components for what
renders across features, _lib for what does not render.

globals.css and favicon.ico stay at the top level deliberately. The
favicon is a Next.js metadata file, and globals.css is the root layout's
stylesheet, which Next.js convention places beside layout.tsx; moving it
into _lib would be tidier by the letter of the rule and stranger to
every Next.js developer who looks for it. The rule in AGENTS.md names
that exception rather than pretending it does not exist.

Worth knowing for the next move of this kind: `tsc --noEmit` was clean
and every `from` specifier correct while three test files were still
broken. `vi.mock("./actions")` takes a path as a string, which the
compiler never checks and a grep for `from "..."` never finds. Only
running the suite caught it.

No behaviour change. 428 tests across 29 files, the same as before, which
for a relocation is the proof that nothing fell out of collection scope.

Closes #144

Entire-Checkpoint: 01M04HX32GCKDSRK567HSATGZE
