+++ CHANGELOG.md
@@ -30,6 +30,7 @@ This project follows a lightweight, human-curated changelog. Keep the newest cha
 
 ### Changed
 
+- The `app/` directory now holds route files and nothing else. It had grown to 23 files of which 15 were not routes — the server action, the cache, the logger, the type definitions and the rest sat at the same level as `page.tsx` and `layout.tsx`, so listing the directory that defines what the app serves mostly showed things it does not serve. Those modules moved to `app/_lib/`, completing the rule the previous two entries started: a folder under `app/` that is not a route starts with an underscore, components live beside the feature that uses them, and everything else is a module in `app/_lib/`. No behaviour changed and no route was added or removed.
 - Components now have one home instead of three. `FirstCommitDisplay` sat alone in a root-level `components/` directory, the four search components sat in `app/_home/`, and the analytics wrapper sat loose in `app/`, with no rule saying which belonged where — so `SearchResults` imported one sibling across an alias boundary and the rest relatively, for no reason a reader could infer. A component now lives beside the feature that uses it, and `app/_components/` holds the ones used by something other than a single feature. The root `components/` directory is gone, and with it the last collection root the unit tests reached outside `app/` and `scripts/`. No behaviour changed; the suite collects the same 428 tests it did before the move.
 - The browser suite now runs on a 320px mobile viewport and on WebKit as well as Desktop Chrome. Only Chrome ran before, so nothing automatically checked how the app behaves at phone width or in the engine behind Safari — layout, focus, navigation, and clipboard were all assumed to work the same everywhere. The two added projects run a selected set of journeys rather than the whole suite, so coverage grows by a few runs instead of tripling. Traces are also retained on failure now: the suite asked for them on the first retry, but retries are off, so no trace was ever captured for any failure in any browser.
 - `AGENTS.md` now carries the agent-rules block that Next.js generates, instead of having it reappear after every browser suite run. `next dev` writes the block whenever its exact text is missing, and the browser suite boots `next dev`, so each full validation run left a modified `AGENTS.md` that someone had to notice and strip — it reached a commit by accident once. The generator offers no way to turn this off, and it skips writing once the block is present, so committing it is what actually stops it. A note above the block records that it is generated, that edits between its markers are overwritten on the next run, and that it cannot move to `CLAUDE.md`, which the gate requires to stay a byte-for-byte pointer.
