+++ app/_home/SearchResults.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { describe, expect, it, vi } from "vitest";
-import type { CommitInfo } from "../commitTypes";
+import type { CommitInfo } from "../_lib/commitTypes";
 import SearchResults from "./SearchResults";
 
 const commits: CommitInfo[] = [
